Fixed basic pause and play

+++ app/src/main/java/com/app/bissudroid/musify/models/Songs.kt
@@ -8,7 +8,7 @@ package com.app.bissudroid.musify.models
       var songDuration: String,
       var songAlbum: String,
       var songPath: String,
-      var albumid: Long,
+      var albumid: Int,
      var mediaId: Long
 
   ) {
